+++ lib/sycamore/sycamore/docset.py
@@ -3,6 +3,7 @@ import logging
 import pprint
 import sys
 from typing import Callable, Optional, Any, Iterable, Type, Union
+from dateutil import parser
 
 from sycamore.context import Context
 from sycamore.data import Document, Element, MetadataDocument
@@ -853,6 +854,84 @@ class DocSet:
         docset = docset.filter(lambda doc: threshold_filter(doc, threshold), **resource_args)
 
         return docset
+    
+    def match_filter(self, query: Any, field: str, ignore_case: bool = True, **resource_args) -> "DocSet":
+        """
+        Only keep documents that match the query on the specified field.
+        Performs substring matching for strings.
+
+        Args:
+            query: Query to match for.
+            field: Document field that is used for filtering.
+            ignore_case: Determines case sensitivity for strings.
+
+        Returns:
+            A filtered DocSet.
+        """
+
+        def process_doc(doc: Document) -> bool:
+            value = doc.field_to_value(field)
+
+            # substring matching
+            if isinstance(query, str) or isinstance(value, str):
+                query_str = str(query)
+                value_str = str(value)
+                if ignore_case:
+                    value_str = value_str.lower()
+                    query_str = query_str.lower()
+
+                return query_str in value_str
+
+            # For non-string types, check exact match
+            return query == value
+
+        return self.filter(process_doc, **resource_args)
+    
+    def range_filter(self, field: str, start: Optional[Any] = None, end: Optional[Any] = None, 
+                     date: Optional[bool] = False, **resource_args) -> "DocSet":
+        """
+        Only keep documents for which the value of the
+        specified field is within the start:end range.
+
+        Args:
+            field: Document field to filter based on.
+            start: Value for start of range.
+            end: Value for end of range.
+            date: Indicates whether start:end is a date range.
+
+        Returns:
+            A filtered DocSet.
+        """
+
+        def process_doc(doc: Document) -> bool:
+            value = doc.field_to_value(field)
+
+            if date:
+                if not isinstance(value, str):
+                    raise ValueError("value must be a string for date filtering")
+                value_comp = parser.parse(value).replace(tzinfo=None)
+                if start and not isinstance(start, str):
+                    raise ValueError("start must be a string for date filtering")
+                start_comp = parser.parse(start).replace(tzinfo=None) if start else None
+                if end and not isinstance(end, str):
+                    raise ValueError("end must be a string for date filtering")
+                end_comp = parser.parse(end).replace(tzinfo=None) if end else None
+            else:
+                value_comp = value
+                start_comp = start
+                end_comp = end
+
+            if start_comp is None:
+                if end_comp is None:
+                    raise ValueError("At least one of start or end must be specified")
+                return value_comp <= end_comp
+            if end_comp is None:
+                if start_comp is None:
+                    raise ValueError("At least one of start or end must be specified")
+                return value_comp >= start_comp
+            return value_comp >= start_comp and value_comp <= end_comp
+
+        return self.filter(process_doc, **resource_args)
 
     def map_batch(
         self,
