+++ lib/sycamore/sycamore/query/execution/operations.py
@@ -44,120 +44,6 @@ SC_ASSIGN_GROUPS_PROMPT = """Categorize the database entry you are given corresp
 into one of the following groups: "{groups}". Perform your best work to assign the group. Return
 ONLY the string corresponding to the selected group. Here is the database entry you will use: """
 
-
-def convert_string_to_date(date_string: str) -> datetime:
-    """
-    Creates datetime object given a date string.
-
-    Args:
-        date_string: The string that contains the date in any format.
-
-    Returns:
-        A datetime object (without timezone info).
-    """
-    return parser.parse(date_string).replace(tzinfo=None)
-
-
-def match_filter_operation(doc: Document, query: Any, field: str, ignore_case: bool = True) -> bool:
-    """
-    Only keep documents that match the query on the specified field.
-    Performs substring matching for strings.
-
-    Args:
-        doc: Document to filter.
-        query: Query to match for.
-        field: Document field that is used for filtering.
-        ignore_case: Determines case sensitivity for strings.
-
-    Returns:
-        A filtered document.
-
-    Example:
-        .. code-block:: python
-
-        def wrapper(doc: Document) -> bool:
-            query = "Cessna"
-            field = "properties.entity.aircraft"
-            return match_filter_operation(doc, query, field)
-
-        docset = docset.filter(wrapper)
-    """
-    value = doc.field_to_value(field)
-
-    # substring matching
-    if isinstance(query, str) or isinstance(value, str):
-        query = str(query)
-        value = str(value)
-        if ignore_case:
-            value = value.lower()
-            query = query.lower()
-
-        return query in value
-
-    # if not string, exact match
-    return query == value
-
-
-def range_filter_operation(
-    doc: Document,
-    field: str,
-    start: Optional[Any] = None,
-    end: Optional[Any] = None,
-    date: Optional[bool] = False,
-) -> bool:
-    """
-    Only keep documents for which the value of the
-    specified field is within the start:end range.
-
-    Args:
-        doc: Document to filter.
-        field: Document field to filter based on.
-        start: Value for start of range.
-        end: Value for end of range.
-        date: Indicates whether start:end is a date range.
-
-    Returns:
-        A filtered document.
-
-    Example:
-        .. code-block:: python
-
-        def wrapper(doc: Document) -> bool:
-            field = "properties.date"
-            start = "July 1, 2020"
-            end = "July 30, 2020"
-            return range_filter_operation(doc, field, start, end, True)
-
-        docset = docset.filter(wrapper)
-    """
-    value = doc.field_to_value(field)
-
-    if date:
-        if not isinstance(value, str):
-            raise ValueError("value must be a string for date filtering")
-        value_comp = convert_string_to_date(value)
-        if start and not isinstance(start, str):
-            raise ValueError("start must be a string for date filtering")
-        start_comp = convert_string_to_date(start) if start else None
-        if end and not isinstance(end, str):
-            raise ValueError("end must be a string for date filtering")
-        end_comp = convert_string_to_date(end) if end else None
-    else:
-        value_comp = value
-        start_comp = start
-        end_comp = end
-
-    if start_comp is None:
-        if end_comp is None:
-            raise ValueError("At least one of start or end must be specified")
-        return value_comp <= end_comp
-    if end_comp is None:
-        if start_comp is None:
-            raise ValueError("At least one of start or end must be specified")
-        return value_comp >= start_comp
-    return value_comp >= start_comp and value_comp <= end_comp
-
-
 def math_operation(val1: int, val2: int, operator: str) -> Union[int, float]:
     """
     Basic arithmetic operations on integers.
