+++ lib/sycamore/sycamore/tests/unit/query/test_operations.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 from typing import Callable, Dict, List, Any, Optional
 
 import pytest
@@ -8,12 +7,9 @@ from sycamore.data import Document
 from sycamore.docset import DocSet
 from sycamore.llms import LLM
 from sycamore.query.execution.operations import (
-    convert_string_to_date,
     join_operation,
     llm_generate_operation,
-    match_filter_operation,
     math_operation,
-    range_filter_operation,
     semantic_cluster,
     top_k_operation,
     SC_FORM_GROUPS_PROMPT,
@@ -82,128 +78,6 @@ class TestOperations:
         return generate_docset(
             {"text_representation": ["1", "2", "one", "two", "1", "3"], "parent_id": [8, 1, 11, 17, 13, 5]},
         )
-
-    def test_match_filter_number(self, words_and_ids_docset):
-        query = 3
-        filtered_docset = words_and_ids_docset.filter(
-            lambda doc: match_filter_operation(doc, query=query, field="doc_id")
-        )
-
-        assert filtered_docset.count() == 2
-        for doc in filtered_docset.take():
-            assert doc.doc_id == 3
-
-    def test_match_filter_string(self, words_and_ids_docset):
-
-        query = "sub"
-        filtered_docset = words_and_ids_docset.filter(
-            lambda doc: match_filter_operation(doc, query=query, field="text_representation")
-        )
-
-        assert filtered_docset.count() == 3
-
-        filtered_texts = []
-        for doc in filtered_docset.take():
-            filtered_texts.append(doc.text_representation)
-        assert filtered_texts == ["submarine", "unSubtle", "Sub"]
-
-    def test_match_filter_string_case_sensititve(self, words_and_ids_docset):
-
-        query = "sub"
-        filtered_docset = words_and_ids_docset.filter(
-            lambda doc: match_filter_operation(doc, query=query, field="text_representation", ignore_case=False)
-        )
-
-        assert filtered_docset.count() == 1
-
-        filtered_texts = []
-        for doc in filtered_docset.take():
-            filtered_texts.append(doc.text_representation)
-        assert filtered_texts == ["submarine"]
-
-    def test_range_filter_number(self, words_and_ids_docset):
-        start, end = 2, 4
-        filtered_docset = words_and_ids_docset.filter(
-            lambda doc: range_filter_operation(doc, field="doc_id", start=start, end=end)
-        )
-
-        assert filtered_docset.count() == 4
-
-        filtered_ids = []
-        for doc in filtered_docset.take():
-            filtered_ids.append(doc.doc_id)
-        assert filtered_ids == [3, 3, 2, 4]
-
-    def test_range_filter_one_sided(self, words_and_ids_docset):
-        start = 5
-        filtered_docset = words_and_ids_docset.filter(
-            lambda doc: range_filter_operation(doc, field="doc_id", start=start, end=None)
-        )
-
-        assert filtered_docset.count() == 4
-
-        filtered_ids = []
-        for doc in filtered_docset.take():
-            filtered_ids.append(doc.doc_id)
-        assert filtered_ids == [5, 9, 6, 7]
-
-        end = 5
-        filtered_docset = words_and_ids_docset.filter(
-            lambda doc: range_filter_operation(doc, field="doc_id", start=None, end=end)
-        )
-
-        assert filtered_docset.count() == 6
-
-        filtered_ids = []
-        for doc in filtered_docset.take():
-            filtered_ids.append(doc.doc_id)
-        assert filtered_ids == [1, 3, 5, 3, 2, 4]
-
-    # only works if the entire field is strings
-    def test_range_filter_string(self, generate_docset):
-        start, end = "b", "t"
-        docset = generate_docset(
-            {"text_representation": ["a", "b", "bBc", "abc", "lmnop", "qq", "edgar", "", "t", "tense"]}
-        )
-        filtered_docset = docset.filter(
-            lambda doc: range_filter_operation(doc, field="text_representation", start=start, end=end)
-        )
-        assert filtered_docset.count() == 6
-
-        filtered_ids = []
-        for doc in filtered_docset.take():
-            filtered_ids.append(doc.text_representation)
-        assert filtered_ids == ["b", "bBc", "lmnop", "qq", "edgar", "t"]
-
-    # only allow date formats recognized by DateUtil parser
-    # e.g. "December 31, 2022, 12:30 Local" will not work
-    def test_range_filter_date(self, generate_docset):
-        start, end = "01-01-2022", "December 31, 2022"
-        docset = generate_docset(
-            {
-                "text_representation": [
-                    "January 1, 2022",
-                    "2/4/20",
-                    "2022-05-04",
-                    "January 14, 2023",
-                    "2023-01-29T12:30:00Z",
-                    "12/12/2023",
-                    "September 19, 2022",
-                    "2022-06-07T03:47:00Z",
-                    "April 15, 2023",
-                ]
-            }
-        )
-        filtered_docset = docset.filter(
-            lambda doc: range_filter_operation(doc, field="text_representation", start=start, end=end, date=True)
-        )
-
-        assert filtered_docset.count() == 4
-
-        filtered_ids = []
-        for doc in filtered_docset.take():
-            filtered_ids.append(doc.text_representation)
-        assert filtered_ids == ["January 1, 2022", "2022-05-04", "September 19, 2022", "2022-06-07T03:47:00Z"]
 
     # LLM Generate
     def test_llm_generate(words_and_ids_docset):
@@ -304,35 +178,35 @@ class TestOperations:
             elif doc.text_representation == "3" or doc.text_representation == "three":
                 assert doc.properties["_autogen_ClusterAssignment"] == "group3"
 
-    def test_convert_string_to_date(self):
-        date_string = "2024-07-21"
-        expected_date = datetime(2024, 7, 21)
-        assert convert_string_to_date(date_string) == expected_date
+    # def test_convert_string_to_date(self):
+    #     date_string = "2024-07-21"
+    #     expected_date = datetime(2024, 7, 21)
+    #     assert convert_string_to_date(date_string) == expected_date
 
-        date_string = "2024-07-21 14:30:00"
-        expected_date = datetime(2024, 7, 21, 14, 30, 0)
-        assert convert_string_to_date(date_string) == expected_date
+    #     date_string = "2024-07-21 14:30:00"
+    #     expected_date = datetime(2024, 7, 21, 14, 30, 0)
+    #     assert convert_string_to_date(date_string) == expected_date
 
-        date_string = "2024-07-21T14:30:00+00:00"
-        expected_date = datetime(2024, 7, 21, 14, 30, 0)
-        assert convert_string_to_date(date_string) == expected_date
+    #     date_string = "2024-07-21T14:30:00+00:00"
+    #     expected_date = datetime(2024, 7, 21, 14, 30, 0)
+    #     assert convert_string_to_date(date_string) == expected_date
 
-        date_string = "21st July 2024"
-        expected_date = datetime(2024, 7, 21)
-        assert convert_string_to_date(date_string) == expected_date
+    #     date_string = "21st July 2024"
+    #     expected_date = datetime(2024, 7, 21)
+    #     assert convert_string_to_date(date_string) == expected_date
 
-        date_string = "Not a date"
-        with pytest.raises(ValueError):
-            convert_string_to_date(date_string)
+    #     date_string = "Not a date"
+    #     with pytest.raises(ValueError):
+    #         convert_string_to_date(date_string)
 
-        date_string = "2020-02-29"
-        expected_date = datetime(2020, 2, 29)
-        assert convert_string_to_date(date_string) == expected_date
+    #     date_string = "2020-02-29"
+    #     expected_date = datetime(2020, 2, 29)
+    #     assert convert_string_to_date(date_string) == expected_date
 
-        date_string = "2023-12-31 23:59:59"
-        expected_date = datetime(2023, 12, 31, 23, 59, 59)
-        assert convert_string_to_date(date_string) == expected_date
+    #     date_string = "2023-12-31 23:59:59"
+    #     expected_date = datetime(2023, 12, 31, 23, 59, 59)
+    #     assert convert_string_to_date(date_string) == expected_date
 
-        date_string = "2024-07-21T14:30:00.123456"
-        expected_date = datetime(2024, 7, 21, 14, 30, 0, 123456)
-        assert convert_string_to_date(date_string) == expected_date
+    #     date_string = "2024-07-21T14:30:00.123456"
+    #     expected_date = datetime(2024, 7, 21, 14, 30, 0, 123456)
+    #     assert convert_string_to_date(date_string) == expected_date
